distributed version WIP

+++ gesv.h
@@ -5,12 +5,12 @@
 
 // General Matrix Solve Vector "scalaire" (solve Ax=b avec b vecteur)
 void tdp_dgesv(const CBLAS_ORDER order, const CBLAS_TRANSPOSE TransA,
-               const int N, double *A, const int lda,
-               double *X, const int incX, const int block_size);
+               const int64_t N, double *A, const int64_t lda,
+               double *X, const int64_t incX, const int64_t block_size);
 
 // general matrix solve vector "bloc" (solve Ax=b avec b vecteur)
 void tdp_dgesv2(const CBLAS_ORDER order, const CBLAS_TRANSPOSE TransA,
-                const int N, double *A, const int lda,
-                double *X, const int incX);
+                const int64_t N, double *A, const int64_t lda,
+                double *X, const int64_t incX);
 
 #endif // TDP_GESV_H
