+++ util.h
@@ -38,28 +38,28 @@
 #define MM256_FMADD_PD(a, b, c) _mm256_add_pd(_mm256_mul_pd(a, b), c)
 #endif
 
-double *tdp_matrix_new(int m/*rows*/, int n/*columns*/);
-double *tdp_avx256_aligned_matrix_new(int m/*rows*/, int n/*columns*/);
-void tdp_matrix_zero(int m/*rows*/, int n/*columns*/, double *mat);
-void tdp_matrix_one(int m/*row*/, int n/*column*/,
-                    double value, double *mat, int lda/*leading dimension*/);
-void tdp_matrix_fill(int m/*row*/, int n/*column*/,
-                     double value, double *mat, int lda/*leading dimension*/);
+double *tdp_matrix_new(int64_t m/*rows*/, int64_t n/*columns*/);
+double *tdp_avx256_aligned_matrix_new(int64_t m/*rows*/, int64_t n/*columns*/);
+void tdp_matrix_zero(int64_t m/*rows*/, int64_t n/*columns*/, double *mat);
+void tdp_matrix_one(int64_t m/*row*/, int64_t n/*column*/,
+                    double value, double *mat, int64_t lda/*leading dimension*/);
+void tdp_matrix_fill(int64_t m/*row*/, int64_t n/*column*/,
+                     double value, double *mat, int64_t lda/*leading dimension*/);
 
-void tdp_matrix_print(int m/*row*/, int n/*column*/,
-                      double *mat, int lda/*leading dimension*/,
+void tdp_matrix_print(int64_t m/*row*/, int64_t n/*column*/,
+                      double *mat, int64_t lda/*leading dimension*/,
                       FILE *outstream);
-void tdp_matrix_rand(int m/*rows*/, int n/*columns*/,
+void tdp_matrix_rand(int64_t m/*rows*/, int64_t n/*columns*/,
                      double *mat, double min, double max);
 
-double *tdp_vector_new(int m);
-void tdp_vector_rand(int m, double min, double max, double *v);
-void tdp_vector_one(int m, double value, double *v);
-void tdp_vector_zero(int m, double *v);
-void tdp_vector_print(int m, double *v, FILE *out);
+double *tdp_vector_new(int64_t m);
+void tdp_vector_rand(int64_t m, double min, double max, double *v);
+void tdp_vector_one(int64_t m, double value, double *v);
+void tdp_vector_zero(int64_t m, double *v);
+void tdp_vector_print(int64_t m, double *v, FILE *out);
 
 void tdp_print_cache_size(void);
-uint64_t tdp_get_cache_size(int id);
+uint64_t tdp_get_cache_size(int64_t id);
 double *tdp_cache_garbage(void);
 
 #endif // TDP_UTIL_H
